// src/types/type.ts
export interface User {
  id?: string
  number?: string
  password?: string
  name?: string
  total?: number
  count?: number
  teacherId?: string
  teacherName?: string
  description?: string
  insertTime?: string
  selectTime?: string
}

export interface ResultVO {
  code: number
  message?: string
  data?: any
}